Changed the fix to solve owin/iis problem
Hey
I encountered the same problem - I have owin app hosted in IIS, and the out of the box example didn't worked because of vistual path problem. Tried the fix in the readme.md, but SwaggerSpecConfig was not found. I looked in the issues, and saw the fixed I used here. It solved the swagger endpoint, but the UI is not working from some reason. Guess this is something worth checking.
The issue I am talking about is issue number 593.
Good Day,
Omer